requirements.txt
Pillow==9.2.0
pyinstaller==6.8.0
pyperclip==1.8.2
screeninfo==0.8.1
